requirements.txt
absl-py==2.1.0
accelerate==0.29.2
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.8
anyascii==0.3.2
anyio==4.3.0
appdirs==1.4.4
appnope @ file:///Users/ktietz/ci_310/appnope_1643965056645/work
argbind==0.3.7
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-timeout==4.0.3
attrs==23.2.0
audiolm-pytorch==1.1.4
audioread==3.0.1
Babel==2.14.0
bangla==0.0.2
beartype==0.18.2
bitarray==2.9.2
blinker==1.7.0
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.6
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comm @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_3doui0bmzb/croot/comm_1709322861485/work
confection==0.1.4
contourpy==1.2.1
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
dateparser==1.1.8
debugpy @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_563_nwtkoc/croot/debugpy_1690905063850/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
descript-audio-codec==1.0.0
descript-audiotools==0.7.2
docker-pycreds==0.4.0
docopt==0.6.2
docstring_parser==0.16
einops==0.7.0
einx==0.2.0
ema-pytorch==0.4.5
encodec==0.1.1
exceptiongroup @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_b2258scr33/croot/exceptiongroup_1706031391815/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fairseq==0.12.2
fastapi==0.110.1
ffmpy==0.3.2
filelock==3.13.1
fire==0.6.0
Flask==3.0.3
flatten-dict==0.4.2
fonttools==4.51.0
frozendict==2.4.1
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
g2pkk==0.1.2
gateloop-transformer==0.2.4
gitdb==4.0.11
GitPython==3.1.43
gradio==4.26.0
gradio_client==0.15.1
grpcio==1.62.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
gruut-lang-es==2.0.0
gruut-lang-fr==2.0.2
h11==0.14.0
hangul-romanize==0.1.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
hydra-core==1.0.7
idna==3.6
importlib_resources==6.4.0
inflect==7.2.0
ipykernel @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_f428_5tjvx/croot/ipykernel_1705933835534/work
ipython @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_a1tmxj9b4u/croot/ipython_1704833016119/work
ipywidgets==8.1.2
itsdangerous==2.1.2
jamo==0.4.1
jedi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/croot-f1t6hma6/jedi_1644315882177/work
jieba==0.42.1
Jinja2==3.1.3
joblib==1.4.0
jsonlines==1.2.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
julius==0.2.7
jupyter_client @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_58w2siozyz/croot/jupyter_client_1699455907045/work
jupyter_core @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_782yoyc_98/croot/jupyter_core_1698937318631/work
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.4
librosa==0.10.0
lion-pytorch==0.1.4
llvmlite==0.42.0
local-attention==1.9.0
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.3
matplotlib==3.8.4
matplotlib-inline @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_f6fdc0hldi/croots/recipe/matplotlib-inline_1662014472341/work
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.2.1
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
nest-asyncio @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_310vb5e2a0/croot/nest-asyncio_1708532678212/work
networkx==2.8.8
nltk==3.8.1
num2words==0.5.13
numba==0.59.1
numpy==1.22.0
omegaconf==2.0.6
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
orjson==3.10.0
packaging @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_a6lqg7at4g/croot/packaging_1710807410750/work
pandas==1.5.3
parler_tts @ git+https://github.com/huggingface/parler-tts.git@10016fb0300c0dc31a0fb70e26f3affee7b62f16
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
Pillow==9.3.0
platformdirs @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_a8u4fy8k9o/croot/platformdirs_1692205661656/work
pooch==1.8.1
portalocker==2.8.2
preshed==3.0.9
progressbar==2.5
prompt-toolkit @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_c63v4kqjzr/croot/prompt-toolkit_1704404354115/work
protobuf==3.19.6
psutil @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_1310b568-21f4-4cb0-b0e3-2f3d31e39728k9coaga5/croots/recipe/psutil_1656431280844/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_29bs9f_dh9/croot/pygments_1684279974747/work
pyloudnorm==0.1.1
pynndescent==0.5.12
pyparsing==3.1.2
pypinyin==0.51.0
pysbd==0.3.4
pystoi==0.4.1
python-crfsuite==0.9.10
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-multipart==0.0.9
pytorch-warmup==0.1.1
pytz==2024.1
PyYAML==6.0.1
pyzmq @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_43pxpbos3z/croot/pyzmq_1705605108344/work
randomname==0.2.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rotary-embedding-torch==0.5.3
rpds-py==0.18.0
ruff==0.3.7
sacrebleu==2.4.1
sacremoses==0.1.1
safetensors==0.4.2
scikit-learn==1.4.2
scipy==1.11.4
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==1.45.0
setproctitle==1.3.3
shellingham==1.5.4
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.4.6
soundfile==0.12.1
sox==1.5.0
soxr==0.3.7
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
starlette==0.37.2
SudachiDict-core==20240409
SudachiPy==0.6.8
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
thinc==8.2.3
threadpoolctl==3.4.0
tiktoken==0.6.0
tokenizers==0.15.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.4.0.dev20240410
torch-stoi==0.2.1
torchaudio==2.2.0.dev20240410
torchvision==0.19.0.dev20240410
tornado @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_3a5nrn2jeh/croot/tornado_1696936974091/work
tortoise-tts @ file:///Users/xuhuizhan/tortoise-tts
tqdm==4.66.2
trainer==0.0.36
traitlets @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_6301rd5qbe/croot/traitlets_1671143894285/work
transformers==4.39.3
TTS==0.22.0
typeguard==4.2.1
typer==0.9.4
typing_extensions==4.11.0
tzlocal==5.2
umap-learn==0.5.6
Unidecode==1.3.8
urllib3==2.2.1
uvicorn==0.29.0
vector_quantize_pytorch==1.14.6
wandb==0.16.6
wasabi==1.1.2
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
weasel==0.3.4
websockets==11.0.3
Werkzeug==3.0.2
widgetsnbextension==4.0.10
yarl==1.9.4
